+++ app/home/route.ts
@@ -2,15 +2,15 @@ import { action } from '@ember-decorators/object';
 import { service } from '@ember-decorators/service';
 import Route from '@ember/routing/route';
 import Ember from 'ember';
-import Analytics from 'ember-osf-web/services/analytics';
 import Session from 'ember-simple-auth/services/session';
+
+import Analytics from 'ember-osf-web/services/analytics';
 
 export default class Home extends Route.extend({
     async beforeModel(this: Home, transition: Ember.Transition) {
         await this._super(transition);
 
         if (this.get('session').get('isAuthenticated')) {
-            transition.abort();
             this.transitionTo('dashboard');
         }
     },