+++ app/services/current-user.ts
@@ -2,20 +2,16 @@ import { computed } from '@ember-decorators/object';
 import { alias } from '@ember-decorators/object/computed';
 import { service } from '@ember-decorators/service';
 import Service from '@ember/service';
-import { task } from 'ember-concurrency';
 import DS from 'ember-data';
-import Features from 'ember-feature-flags';
 import config from 'ember-get-config';
 import Session from 'ember-simple-auth/services/session';
 import RSVP from 'rsvp';
 
 import User from 'ember-osf-web/models/user';
-import authenticatedAJAX from 'ember-osf-web/utils/ajax-helpers';
 
 const {
     OSF: {
         url: osfUrl,
-        apiUrl,
     },
 } = config;
 
@@ -39,9 +35,7 @@ enum AuthRoute {
 export default class CurrentUserService extends Service {
     @service store!: DS.Store;
     @service session!: Session;
-    @service features!: Features;
 
-    waffleLoaded = false;
     showTosConsentBanner = false;
 
     /**
@@ -50,7 +44,7 @@ export default class CurrentUserService extends Service {
     @alias('session.data.authenticated.id') currentUserId: string | undefined;
 
     /**
-     * Return an observable promise proxy for the currently logged in user. If no user is logged in, resolves to null.
+     * Return the currently logged-in user, or null if not logged in.
      */
     @computed('currentUserId')
     get user(this: CurrentUserService): User | null {
@@ -61,48 +55,9 @@ export default class CurrentUserService extends Service {
         return null;
     }
 
-    /**
-     * Fetch waffle flags for the current user and set the corresponding feature flags.
-     */
-    setWaffle = task(function *(this: CurrentUserService) {
-        const { data } = yield authenticatedAJAX({
-            url: `${apiUrl}/v2/_waffle/`,
-        });
-
-        // eslint-disable-next-line no-restricted-globals
-        interface Feature { attributes: { name: string; active: boolean; }; }
-
-        this.get('features').setup(
-            data.reduce((acc: object, { attributes: { name, active } }: Feature) => ({ ...acc, [name]: active }), {}),
-        );
-
-        this.set('waffleLoaded', true);
-    }).drop();
-
     constructor() {
         super();
-
-        function performSetWaffle(this: CurrentUserService) {
-            this.get('setWaffle').perform();
-        }
-
-        this.session.on('authenticationSucceeded', this, performSetWaffle);
         this.session.on('invalidationSucceeded', this, this.logout);
-    }
-
-    /**
-     * Check whether the given waffle/feature flag is enabled.
-     */
-    async getWaffle(this: CurrentUserService, feature: string): Promise<boolean> {
-        const setWaffle = this.get('setWaffle');
-
-        if (setWaffle.isRunning) {
-            await setWaffle.last;
-        } else if (!this.waffleLoaded) {
-            await setWaffle.perform();
-        }
-
-        return this.features.isEnabled(feature);
     }
 
     /**