+++ app/support/route.ts
@@ -1,6 +1,7 @@
 import { action } from '@ember-decorators/object';
 import { service } from '@ember-decorators/service';
 import Route from '@ember/routing/route';
+
 import Analytics from 'ember-osf-web/services/analytics';
 
 export default class Support extends Route {