+++ app/utils/transition-target-url.ts
@@ -1,23 +1,26 @@
-import { Registry as ServiceRegistry } from '@ember/service';
-
 interface Transition {
     targetName: string;
     params: any[];
     queryParams: object;
+    router: {
+        generate(
+            handlerName: string,
+            ...params: any[],
+        ): string;
+    };
 }
 
 /**
  * Get the URL (path and query string) that the given transition will resolve to,
  * using the given router.
  */
-export default function transitionTarget(
+export default function transitionTargetURL(
     transition: Transition,
-    router: ServiceRegistry['router'],
 ) {
     const params = Object.values(transition.params).filter(
         param => Object.values(param).length,
     );
-    return router.urlFor(
+    return transition.router.generate(
         transition.targetName,
         ...params,
         { queryParams: transition.queryParams },