+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2018-07-13
+### Added
+- Models:
+    - `banner` - used to fetch `/_/banners/current/` from the API
+- Components:
+    - `scheduled-banner` - display the "current" banner on the landing page(s)
+
+### Changed
+- refactored feature flags to be loaded from API base (`/v2`)
+- check route feature flags in `Router._doTransition` instead of `Router.willTransition`
+
+### Removed
+- Feature Flags:
+    - `ember_support_page`
+    - `ember_home_page`
+
 ## [0.5.2] - 2018-07-11
 ### Fixed
 - Lazy engine title stealing